Recast: fix regression causing tile size slider to not work
Regression was caused in commit f95eb13ab30c78c03c38aa20492389d0fdf39a7f. The m_tileSize member was moved to the base class as this allows us to run the navmesh type selector in the base class which is subclassed by all the other 3 editors. The member wasn't removed from the tilemesh editor class causing it to shadow the base class one.