Recast: disable depth testing by default for navmesh
The depth testing is useful when rendering the navmesh without a background (e.g. disabling the rendering of the input geom). But when rendering with the geometry, it should be disabled as otherwise some data will overlap and become invisible (the off-mesh links or poly verts, or boundaries when looking from a tilted perspective).